+++ src/main/java/com/portailconge/portail_conge/repository/DepartementRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepartementRepository extends JpaRepository<Departement, Long> {
+public interface DepartementRepository extends JpaRepository<Departement, Integer> {
 }
